requirements.txt
Flask
pymongo
tweepy
facebook-sdk
nltk
